+++ CHANGELOG.md
@@ -12,6 +12,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 <!-- ### Changed -->
 
 - docs(repo): update CONTRIBUTING.md
+- refactor(internal): drop pkg layer
 
 <!-- ### Deprecated -->
