+++ internal/assets/.go-cli.example.yaml
